+++ src/binary/audio/synth/oscillators/hq/phasors/sample_pulse.h
@@ -1,6 +1,6 @@
 #ifndef SAMPLE_PULSE
 #define SAMPLE_PULSE
 
-int sample_pulse(int freq, int t, int duty_cycle);
+double sample_pulse(int freq, int t, int duty_cycle);
 
 #endif
