+++ src/binary/audio/synth/oscillators/hq/phasors/sample_square.h
@@ -1,6 +1,6 @@
 #ifndef SAMPLE_SQUARE
 #define SAMPLE_SQUARE
 
-int sample_square(int freq, int t);
+double sample_square(int freq, int t);
 
 #endif
